+++ service/controller/clusterapi/resource.go
@@ -184,9 +184,14 @@ func New(config Config) ([]resource.Interface, error) {
 			return nil, microerror.Mask(err)
 		}
 	}
+
 	organizationReader := organization.NewNamespaceReader(config.K8sClient.K8sClient(), config.Installation, config.Provider)
+	// This resource creates a the prometheus agent remote write configuration.
+	// This is now managed by the observability-operator when mimir is enabled.
 	var remoteWriteConfigResource resource.Interface
-	{
+	if config.MimirEnabled {
+		remoteWriteConfigResource = noop.New(noop.Config{Logger: config.Logger})
+	} else {
 		c := remotewriteconfig.Config{
 			K8sClient:          config.K8sClient,
 			Logger:             config.Logger,
@@ -208,8 +213,12 @@ func New(config Config) ([]resource.Interface, error) {
 		}
 	}
 
+	// This resource creates a the prometheus agent remote write secret.
+	// This is now managed by the observability-operator when mimir is enabled.
 	var remoteWriteSecretResource resource.Interface
-	{
+	if config.MimirEnabled {
+		remoteWriteSecretResource = noop.New(noop.Config{Logger: config.Logger})
+	} else {
 		c := remotewritesecret.Config{
 			K8sClient:       config.K8sClient,
 			Logger:          config.Logger,
@@ -226,8 +235,11 @@ func New(config Config) ([]resource.Interface, error) {
 		}
 	}
 
+	// This resource is not used in latest observability bundle versions.
 	var remoteWriteAPIEndpointConfigSecretResource resource.Interface
-	{
+	if config.MimirEnabled {
+		remoteWriteAPIEndpointConfigSecretResource = noop.New(noop.Config{Logger: config.Logger})
+	} else {
 		c := remotewriteapiendpointconfigsecret.Config{
 			K8sClient:          config.K8sClient,
 			Logger:             config.Logger,
@@ -395,13 +407,13 @@ func New(config Config) ([]resource.Interface, error) {
 		rbacResource,
 		heartbeatWebhookConfigResource,
 		scrapeConfigResource,
+		alertmanagerWiringResource,
+		prometheusResource,
 		remoteWriteConfigResource,
 		remoteWriteSecretResource,
 		remoteWriteAPIEndpointConfigSecretResource,
 		remoteWriteIngressAuthResource,
 		remoteWriteIngressResource,
-		alertmanagerWiringResource,
-		prometheusResource,
 		verticalPodAutoScalerResource,
 		ingressResource,
 		heartbeatResource,