+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Remove prometheus remote write agent configuration when mimir is enabled.
+
 ## [4.75.0] - 2024-05-13
 
 ### Added